tests/wpt/tests/css/css-typed-om/the-stylepropertymap/properties/gap.html

<!doctype html>
<meta charset="utf-8">
<title>'*-gap' properties</title>
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om/#dom-stylepropertymapreadonly-get">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om/#dom-stylepropertymap-set">
<link rel="help" href="https://drafts.css-houdini.org/css-typed-om/#reify-stylevalue">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../resources/testhelper.js"></script>
<script src="resources/testsuite.js"></script>
<body>
<div id="log"></div>
<script>
'use strict';

function assert_is_equal_with_clamping_percentage(input, result) {
  const percent = input.to('percent').value;

  if (percent < 0)
    assert_style_value_equals(result, new CSSUnitValue(0, 'percent'));
  else
    assert_style_value_equals(result, new CSSUnitValue(percent, 'percent'));
}

for (const prefix of ['column', 'row']) {
  runPropertyTests(`${prefix}-gap`, [
    { syntax: 'normal' },
    {
      syntax: '<length>',
      specified: assert_is_equal_with_range_handling
    },
    {
      syntax: '<percentage>',
      specified: assert_is_equal_with_range_handling,
      computed: assert_is_equal_with_clamping_percentage
    },
  ]);
}

</script>
