third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/gap-expected.txt

This is a testharness.js-based test.
Found 2 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Can set 'column-gap' to a percent: -3.14%
  assert_approx_equals: expected 0 +/- 0.000001 but got -3.140000104904175
[FAIL] Can set 'row-gap' to a percent: -3.14%
  assert_approx_equals: expected 0 +/- 0.000001 but got -3.140000104904175
Harness: the test ran to completion.
